Detect cycle in an undirected graph
